Move examples directory out of lib
The examples directory should be located on top level
