+++ .prettierignore
@@ -1,3 +1,3 @@
 node_modules/
 dist/
-lib/examples/build/
+examples/build/
