+++ examples/create-minimal-sample-sbom.ts
@@ -1,6 +1,6 @@
-import * as sbom from "../spdx-tools";
+import * as sbom from "../lib/spdx-tools";
 
-const sampleSbom = "./lib/examples/resources/minimal-sample.sbom.json";
+const sampleSbom = "./examples/resources/minimal-sample.sbom.json";
 const document = sbom.createDocument(
   "first-document",
   { name: "test creator", type: "Person" },
