+++ examples/package.json
@@ -1,0 +1,1 @@
+{"type": "module"}
