+++ package.json
@@ -30,8 +30,8 @@
   },
   "scripts": {
     "build": "rollup --config rollup.config.ts --configPlugin typescript",
-    "build-examples": "rollup --config lib/examples/rollup.config.ts --configPlugin typescript",
-    "create-sample-sboms": "node lib/examples/build/create-minimal-sample-sbom.js && node lib/examples/build/create-elaborate-sample-sbom.js",
+    "build-examples": "rollup --config examples/rollup.config.ts --configPlugin typescript",
+    "create-sample-sboms": "node examples/build/create-minimal-sample-sbom.js && node examples/build/create-elaborate-sample-sbom.js",
     "lint-check": "eslint -c .eslintrc.js --ext .ts,.js .",
     "lint-fix": "eslint -c .eslintrc.js --fix --ext .ts,.js .",
     "prepare": "husky install && npm run build",
